Safari-611.1.10/LayoutTests/js/dom/exception-line-number-expected.txt

This test makes sure the exception stack has line number information.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is "LayoutTests/js/dom/script-tests/exception-line-number.js:6,LayoutTests/js/dom/script-tests/exception-line-number.js:11"
PASS result is "LayoutTests/js/dom/script-tests/exception-line-number.js:36:15"
PASS successfullyParsed is true

TEST COMPLETE
